# apps/control-plane/src/control_plane/routes/org.py
from __future__ import annotations

from datetime import UTC, datetime
from uuid import UUID  # noqa: TC003 fastapi resolves path param annotations at runtime

from fastapi import APIRouter, HTTPException

from control_plane.authority import ensure_org_role_change
from control_plane.authz import OrgRole, Permission, Scope
from control_plane.compiler import publish_pending
from control_plane.deps import ActorDep, OrgDep, WorkspaceDep, org_scope, require, require_all, workspace_scope
from control_plane.models import AuditLog, Bundle, OrgMembership, RuntimeConfiguration, UsageEvent, User
from control_plane.models.audit import ActivityOut
from control_plane.models.bundle import BundleOut
from control_plane.models.common import PageDep  # noqa: TC001 FastAPI resolves route annotations at runtime
from control_plane.models.common.wire import DeletedOut, Envelope, PageEnvelope
from control_plane.models.management_key import ManagementKeyCreatedOut, ManagementKeyIn  # noqa: TC001 FastAPI resolves route annotations at runtime
from control_plane.models.org_membership import MembershipOut, OrgMemberOut, OrgMembershipIn
from control_plane.models.usage_event import UsageEventOut
from control_plane.models.user import OrgServiceAccountCreatedOut, OrgServiceAccountIn, UserOut
from control_plane.routes.management_keys import issue_management_key

router = APIRouter(prefix="/organizations/{org_id}")


@router.get("/users", tags=["Organization Members"], dependencies=[require("api", org_scope, Permission.members_read)])
async def list_org_users(org_id: OrgDep, page: PageDep) -> PageEnvelope[OrgMemberOut]:
    """List the human users and service accounts that belong to an organization."""
    members = await User.page_members_of(org_id, page)
    roles = await OrgMembership.roles_for_org_users(org_id, tuple(user.id for user in members.items))
    return PageEnvelope.from_slice(
        members.map(
            lambda user: OrgMemberOut(
                user_id=user.id,
                email=user.email,
                name=user.name,
                service_account=user.service_account,
                role=roles[user.id],
                status="member",
                managed=user.managing_org_id == org_id,
            )
        )
    )


@router.put("/users/{user_id}", tags=["Organization Members"], dependencies=[require("api", org_scope, Permission.members_manage)])
async def add_org_user(user_id: UUID, body: OrgMembershipIn, org_id: OrgDep, actor: ActorDep) -> Envelope[MembershipOut]:
    """Add a principal to an organization or update its organization role."""
    user = await User.find_by_id(user_id)
    if user is None:
        raise HTTPException(status_code=404, detail="User not found")
    if user.managing_org_id is not None and user.managing_org_id != org_id:
        raise HTTPException(status_code=409, detail="Organization-managed service accounts cannot join another organization")
    if user.managing_org_id is not None and body.role is OrgRole.owner:
        raise HTTPException(status_code=409, detail="Organization-managed service accounts cannot own an organization")
    membership = await OrgMembership.get((user_id, org_id))
    current = membership.role if membership else None
    await ensure_org_role_change(actor, org_id, current, body.role)
    if membership is None:
        membership = OrgMembership(user_id=user_id, org_id=org_id, role=body.role)
        await membership.save()
    elif membership.role != body.role:
        await membership.change_role(body.role)
    return Envelope(data=MembershipOut(user_id=user_id, org_id=org_id, role=membership.role, status="member"))


@router.delete("/users/{user_id}", tags=["Organization Members"], dependencies=[require("api", org_scope, Permission.members_manage)])
async def remove_org_user(user_id: UUID, org_id: OrgDep, actor: ActorDep) -> Envelope[DeletedOut[str]]:
    """Remove a principal from an organization and its workspaces."""
    membership = await OrgMembership.get((user_id, org_id))
    if membership is None:
        raise HTTPException(status_code=404, detail="User is not a member of this org")
    user = await User.find_by_id(user_id)
    if user is not None and user.managing_org_id == org_id:
        raise HTTPException(status_code=409, detail="Delete an organization-managed service account instead of removing its membership")
    await ensure_org_role_change(actor, org_id, membership.role, OrgRole.member)
    await membership.delete()
    return Envelope(data=DeletedOut.of(f"{user_id}/{org_id}"))


@router.post(
    "/service-accounts",
    tags=["Organization Service Accounts"],
    dependencies=[require_all("api", org_scope, Permission.members_manage, Permission.management_keys_issue)],
)
async def create_org_service_account(
    body: OrgServiceAccountIn,
    org_id: OrgDep,
    actor: ActorDep,
) -> Envelope[OrgServiceAccountCreatedOut]:
    """Create an organization-managed service account and its first management key."""
    await ensure_org_role_change(actor, org_id, None, OrgRole.admin)
    service_account = await User.new_service_account(body.name, managing_org_id=org_id).save()
    membership = await OrgMembership(user_id=service_account.id, org_id=org_id, role=OrgRole.admin).save()
    management_key = await issue_management_key(
        body.management_key,
        actor,
        Scope.org(org_id),
        principal_id=service_account.id,
    )
    return Envelope(
        data=OrgServiceAccountCreatedOut(
            service_account=UserOut.model_validate({**service_account.model_dump(), "org_count": 1}),
            membership=MembershipOut(user_id=service_account.id, org_id=org_id, role=OrgRole(membership.role), status="member"),
            management_key=management_key,
        )
    )


@router.post(
    "/service-accounts/{user_id}/management-keys",
    tags=["Organization Service Accounts"],
    dependencies=[require("api", org_scope, Permission.management_keys_issue)],
)
async def create_org_service_account_management_key(
    user_id: UUID,
    body: ManagementKeyIn,
    org_id: OrgDep,
    actor: ActorDep,
) -> Envelope[ManagementKeyCreatedOut]:
    """Issue a replacement key for an organization-managed service account."""
    service_account = await User.owned_by(org_id, user_id)
    return Envelope(data=await issue_management_key(body, actor, Scope.org(org_id), principal_id=service_account.id))


@router.delete(
    "/service-accounts/{user_id}",
    tags=["Organization Service Accounts"],
    dependencies=[require_all("api", org_scope, Permission.members_manage, Permission.management_keys_revoke)],
)
async def delete_org_service_account(user_id: UUID, org_id: OrgDep, actor: ActorDep) -> Envelope[DeletedOut[UUID]]:
    """Delete an organization-managed service account and its control-plane credentials."""
    service_account = await User.owned_by(org_id, user_id)
    membership = await OrgMembership.get((user_id, org_id))
    if membership is not None:
        await ensure_org_role_change(actor, org_id, membership.role, OrgRole.member)
    if membership is not None:
        await membership.delete()
    await service_account.delete_with_contents()
    return Envelope(data=DeletedOut.of(user_id))


@router.post("/bundles/republish", tags=["Organization Bundles"], dependencies=[require("api", org_scope, Permission.bundles_publish)])
async def republish_bundle(org_id: OrgDep) -> Envelope[BundleOut]:
    """Request a fresh bundle for the organization's current configuration."""
    now = datetime.now(tz=UTC)
    await RuntimeConfiguration.request_republication(org_id)
    published = await publish_pending(now)
    bundle = next(bundle for bundle in published if bundle.org_id == org_id)
    return Envelope(data=BundleOut.model_validate(bundle))


@router.get("/bundles", tags=["Organization Bundles"], dependencies=[require("api", org_scope, Permission.bundles_read)])
async def list_bundles(org_id: OrgDep, page: PageDep) -> PageEnvelope[BundleOut]:
    """List policy bundle metadata for an organization."""
    return PageEnvelope.from_slice(await Bundle.page_for_org(org_id, page), BundleOut)


@router.get("/events", tags=["Organization Usage Events"], dependencies=[require("api", org_scope, Permission.usage_read)])
async def list_org_events(org_id: OrgDep, page: PageDep) -> PageEnvelope[UsageEventOut]:
    """List usage events across an organization with cursor pagination."""
    return PageEnvelope.from_slice(await UsageEvent.for_scope(org_id, None, page), UsageEventOut)


@router.get(
    "/workspaces/{workspace_ref}/events", tags=["Workspace Usage Events"], dependencies=[require("api", workspace_scope, Permission.usage_read)]
)
async def list_workspace_events(workspace: WorkspaceDep, page: PageDep) -> PageEnvelope[UsageEventOut]:
    """List usage events for one workspace with cursor pagination."""
    return PageEnvelope.from_slice(await UsageEvent.for_scope(workspace.org_id, workspace.id, page), UsageEventOut)


@router.get("/activity", tags=["Organization Activity"], dependencies=[require("api", org_scope, Permission.audit_read)])
async def list_activity(org_id: OrgDep, page: PageDep) -> PageEnvelope[ActivityOut]:
    """List the most recent audited changes in an organization."""
    return PageEnvelope.from_slice(await AuditLog.for_org(org_id, page), ActivityOut)