# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: airmux Control Plane API
  description: 'Manage airmux organizations, workspaces, management keys, provider credentials, and data-plane synchronization.


    Organization and workspace targets are part of each URL. Successful responses wrap their result in

    `{"data": ...}`. Authenticate with a control-plane management key or a browser session as documented by

    each operation. Inference keys authenticate model requests at the data plane and are not accepted here.

    '
  version: 0.1.0
paths:
...
  /api/v1/service-accounts:
    post:
      tags:
      - Instance Users
      summary: Create Service Account
      description: 'Create a machine principal with an optional instance role.


        Required permission: `principals.manage`.'
      operationId: create_service_account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceAccountIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_UserOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      security:
      - ManagementKey: []
      - SessionCookie: []
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - principals.manage
  /api/v1/service-accounts/{user_id}/management-keys:
    post:
      tags:
      - Instance Users
      summary: Create Instance Service Account Management Key
      description: 'Issue an instance key for an instance-managed service account.


        Required permission: `management-keys.issue`.'
      operationId: create_instance_service_account_management_key
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagementKeyIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ManagementKeyCreatedOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - management-keys.issue
  /api/v1/users/{user_id}:
    get:
      tags:
      - Instance Users
      summary: Get User
      description: 'Return one human user or service account and its organization memberships.


        Required permission: `principals.read`.'
      operationId: get_user
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_UserOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - principals.read
    delete:
      tags:
      - Instance Users
      summary: Delete User
      description: 'Delete a principal and its login identities, sessions, and control-plane management keys.


        Remove organization memberships, personal organizations, and workspaces containing inference

        keys created by this principal before deleting it.


        Required permission: `principals.manage`.'
      operationId: delete_user
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DeletedOut_UUID__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - principals.manage
...
  /api/v1/users:
    get:
      tags:
      - Instance Users
      summary: List Users
      description: 'List human users and service accounts across the instance.


        Required permission: `principals.read`.'
      operationId: list_users
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: service_account
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Service Account
        description: 'Filter by principal type: true for service accounts and false for human users'
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/CursorToken'
          - type: 'null'
          title: Cursor
        description: Opaque continuation token from the previous page
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
        description: Maximum number of results to return
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEnvelope_UserOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - principals.read
...
  /api/v1/organizations/{org_id}:
    patch:
      tags:
      - Organization Settings
      summary: Update Organization
      description: 'Update an organization''s mutable fields.


        Required permission: `organizations.update`.'
      operationId: update_org
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
        description: Organization ID or slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_OrgOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: org_scope
        anyOf:
        - organizations.update
    get:
      tags:
      - Organization Settings
      summary: Get Organization
      description: 'Return one organization.


        Required permission: `organizations.read`.'
      operationId: get_org
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
        description: Organization ID or slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_OrgOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: org_scope
        anyOf:
        - organizations.read
    delete:
      tags:
      - Organization Settings
      summary: Delete Organization
      description: 'Delete an organization and its workspaces, keys, credentials, memberships, and bundles.


        Historical usage events and audit entries are retained.


        Required permission: `organizations.delete`.'
      operationId: delete_org
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
        description: Organization ID or slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DeletedOut_UUID__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: org_scope
        anyOf:
        - organizations.delete
...
  /api/v1/organizations/{org_id}/users:
    get:
      tags:
      - Organization Members
      summary: List Organization Members
      description: 'List the human users and service accounts that belong to an organization.


        Required permission: `members.read`.'
      operationId: list_org_users
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
        description: Organization ID or slug
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/CursorToken'
          - type: 'null'
          title: Cursor
        description: Opaque continuation token from the previous page
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
        description: Maximum number of results to return
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEnvelope_OrgMemberOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: org_scope
        anyOf:
        - members.read
  /api/v1/organizations/{org_id}/users/{user_id}:
    put:
      tags:
      - Organization Members
      summary: Add Organization Member
      description: 'Add a principal to an organization or update its organization role.


        Required permission: `members.manage`.'
      operationId: add_org_user
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
        description: Organization ID or slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgMembershipIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_MembershipOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: org_scope
        anyOf:
        - members.manage
    delete:
      tags:
      - Organization Members
      summary: Remove Organization Member
      description: 'Remove a principal from an organization and its workspaces.


        Required permission: `members.manage`.'
      operationId: remove_org_user
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
        description: Organization ID or slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DeletedOut_str__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: org_scope
        anyOf:
        - members.manage
  /api/v1/organizations/{org_id}/service-accounts:
    post:
      tags:
      - Organization Service Accounts
      summary: Create Organization Service Account
      description: 'Create an organization-managed service account and its first management key.


        Required permissions: `members.manage` and `management-keys.issue`.'
      operationId: create_org_service_account
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
        description: Organization ID or slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgServiceAccountIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_OrgServiceAccountCreatedOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: org_scope
        anyOf:
        - members.manage
      - scope: org_scope
        anyOf:
        - management-keys.issue
  /api/v1/organizations/{org_id}/service-accounts/{user_id}/management-keys:
    post:
      tags:
      - Organization Service Accounts
      summary: Create Org Service Account Management Key
      description: 'Issue a replacement key for an organization-managed service account.


        Required permission: `management-keys.issue`.'
      operationId: create_org_service_account_management_key
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
        description: Organization ID or slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagementKeyIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ManagementKeyCreatedOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: org_scope
        anyOf:
        - management-keys.issue
  /api/v1/organizations/{org_id}/service-accounts/{user_id}:
    delete:
      tags:
      - Organization Service Accounts
      summary: Delete Organization Service Account
      description: 'Delete an organization-managed service account and its control-plane credentials.


        Required permissions: `members.manage` and `management-keys.revoke`.'
      operationId: delete_org_service_account
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
        description: Organization ID or slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DeletedOut_UUID__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: org_scope
        anyOf:
        - members.manage
      - scope: org_scope
        anyOf:
        - management-keys.revoke
  /api/v1/organizations/{org_id}/bundles/republish:
    post:
      tags:
      - Organization Bundles
      summary: Republish Policy Bundle
      description: 'Request a fresh bundle for the organization''s current configuration.


        Required permission: `bundles.publish`.'
      operationId: republish_bundle
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
        description: Organization ID or slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_BundleOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: org_scope
        anyOf:
        - bundles.publish
  /api/v1/organizations/{org_id}/bundles:
    get:
      tags:
      - Organization Bundles
      summary: List Bundles
      description: 'List policy bundle metadata for an organization.


        Required permission: `bundles.read`.'
      operationId: list_bundles
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
        description: Organization ID or slug
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/CursorToken'
          - type: 'null'
          title: Cursor
        description: Opaque continuation token from the previous page
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
        description: Maximum number of results to return
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEnvelope_BundleOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: org_scope
        anyOf:
        - bundles.read
...
components:
  schemas:
...
    CursorToken:
      type: string
      maxLength: 512
      minLength: 1
      pattern: ^[A-Za-z0-9_-]+$
...
    Envelope_DeletedOut_UUID__:
      properties:
        data:
          $ref: '#/components/schemas/DeletedOut_UUID_'
      type: object
      required:
      - data
      title: Envelope[DeletedOut[UUID]]
...
    Envelope_ManagementKeyCreatedOut_:
      properties:
        data:
          $ref: '#/components/schemas/ManagementKeyCreatedOut'
      type: object
      required:
      - data
      title: Envelope[ManagementKeyCreatedOut]
...
    Envelope_UserOut_:
      properties:
        data:
          $ref: '#/components/schemas/UserOut'
      type: object
      required:
      - data
      title: Envelope[UserOut]
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ManagementKeyIn:
      properties:
        permissions:
          items:
            $ref: '#/components/schemas/Permission'
          type: array
          minItems: 1
          title: Permissions
          description: Explicit maximum permissions carried by the key
        label:
          type: string
          maxLength: 80
          minLength: 1
          title: Label
          description: Where this key lives, such as ci, laptop, or data-plane
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
          description: Optional expiration timestamp with a timezone
      additionalProperties: false
      type: object
      required:
      - permissions
      - label
      title: ManagementKeyIn
...
    OrgUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Name
          description: Replacement organization name
      additionalProperties: false
      type: object
      title: OrgUpdate
...
    ServiceAccountIn:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
          description: Display name for the service account
        instance_role:
          anyOf:
          - $ref: '#/components/schemas/InstanceRole'
          - type: 'null'
          description: Optional instance-wide role for the service account
      additionalProperties: false
      type: object
      required:
      - name
      title: ServiceAccountIn
...
  securitySchemes:
    ManagementKey:
      type: http
      description: A control-plane management key using the `sk-cp-` prefix. Inference keys are not accepted by control-plane
        endpoints.
      scheme: bearer
    SessionCookie:
      type: apiKey
      in: cookie
      name: airmux_session
      description: Browser session cookie returned by login or signup. Browser requests must also send `X-Requested-With`.
...